# docs/swagger.yaml
basePath: /v1
definitions:
  entity.Task:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      description:
        example: Task description
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      status:
        allOf:
        - $ref: '#/definitions/entity.TaskStatus'
        example: todo
      title:
        example: My task
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  entity.TaskStatus:
    enum:
    - todo
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - TaskStatusTodo
    - TaskStatusInProgress
    - TaskStatusDone
  entity.Translation:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
      translation:
        example: text for translation
        type: string
    type: object
  entity.TranslationHistory:
    properties:
      history:
        items:
          $ref: '#/definitions/entity.Translation'
        type: array
    type: object
  entity.User:
    properties:
      created_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      updated_at:
        example: "2026-01-01T00:00:00Z"
        type: string
      username:
        example: johndoe
        type: string
    type: object
  v1.CreateTask:
    properties:
      description:
        example: Task description
        maxLength: 1000
        type: string
      title:
        example: My task
        maxLength: 255
        type: string
    required:
    - title
    type: object
  v1.Error:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: invalid request body
        type: string
    type: object
  v1.Login:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  v1.Ok-any:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: ok
        type: string
    type: object
  v1.Ok-entity_Task:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/entity.Task'
      message:
        example: ok
        type: string
    type: object
  v1.Ok-entity_Translation:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/entity.Translation'
      message:
        example: ok
        type: string
    type: object
  v1.Ok-entity_TranslationHistory:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/entity.TranslationHistory'
      message:
        example: ok
        type: string
    type: object
  v1.Ok-entity_User:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/entity.User'
      message:
        example: ok
        type: string
    type: object
  v1.Ok-v1_TaskList:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/v1.TaskList'
      message:
        example: ok
        type: string
    type: object
  v1.Ok-v1_Token:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/v1.Token'
      message:
        example: ok
        type: string
    type: object
  v1.Register:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
      username:
        example: johndoe
        maxLength: 255
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  v1.TaskList:
    properties:
      tasks:
        items:
          $ref: '#/definitions/entity.Task'
        type: array
      total:
        example: 42
        type: integer
    type: object
  v1.Token:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
  v1.TransitionTask:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/entity.TaskStatus'
        enum:
        - todo
        - in_progress
        - done
        example: in_progress
    required:
    - status
    type: object
  v1.Translate:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
    required:
    - destination
    - original
    - source
    type: object
  v1.UpdateTask:
    properties:
      description:
        example: Updated description
        maxLength: 1000
        type: string
      title:
        example: Updated task
        maxLength: 255
        type: string
    required:
    - title
    type: object
host: localhost:8080
info:
  contact: {}
  description: Multi-domain clean architecture template with translation, user, and
    task management
  title: Go Clean Template API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get JWT token
      operationId: login
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-v1_Token'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/v1.Token'
                Message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      summary: Login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      operationId: register
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.Register'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_User'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.User'
                Message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      summary: Register
      tags:
      - auth
  /tasks:
    get:
      description: List tasks for the current user with optional filtering
      operationId: list-tasks
      parameters:
      - description: Filter by status
        enum:
        - todo
        - in_progress
        - done
        in: query
        name: status
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-v1_TaskList'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/v1.TaskList'
                Message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task for the current user
      operationId: create-task
      parameters:
      - description: Task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.CreateTask'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_Task'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.Task'
                Message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Create task
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Delete a task by ID
      operationId: delete-task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-any'
            - properties:
                Code:
                  type: integer
                Data:
                  type: object
                Message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Delete task
      tags:
      - tasks
    get:
      description: Get a task by ID
      operationId: get-task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_Task'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.Task'
                Message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update task title and description
      operationId: update-task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated task data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_Task'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.Task'
                Message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Update task
      tags:
      - tasks
  /tasks/{id}/status:
    patch:
      consumes:
      - application/json
      description: Change task status (todo -> in_progress -> done, or in_progress
        -> todo)
      operationId: transition-task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.TransitionTask'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_Task'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.Task'
                Message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Transition task status
      tags:
      - tasks
  /translation/do-translate:
    post:
      consumes:
      - application/json
      description: Translate a text
      operationId: do-translate
      parameters:
      - description: Set up translation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.Translate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_Translation'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.Translation'
                Message:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Translate
      tags:
      - translation
  /translation/history:
    get:
      description: Show all translation history for current user
      operationId: history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_TranslationHistory'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.TranslationHistory'
                Message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Show history
      tags:
      - translation
  /user/profile:
    get:
      description: Get current user profile
      operationId: profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.Ok-entity_User'
            - properties:
                Code:
                  type: integer
                Data:
                  $ref: '#/definitions/entity.User'
                Message:
                  type: string
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.Error'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - user
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"