# internal/controller/restapi/v1/translation.go
package v1

import (
	"net/http"

	"github.com/gofiber/fiber/v3"
	"github.com/leijux/go-clean-template/internal/controller/restapi/v1/request"
	_ "github.com/leijux/go-clean-template/internal/controller/restapi/v1/response" // for swaggo
	"github.com/leijux/go-clean-template/internal/entity"
)

// @Summary     Show history
// @Description Show all translation history for current user
// @ID          history
// @Tags        translation
// @Produce     json
// @Success     200 {object} response.Ok[entity.TranslationHistory]{Code=int,Message=string,Data=entity.TranslationHistory}
// @Failure     401 {object} response.Error
// @Failure     500 {object} response.Error
// @Security    BearerAuth
// @Router      /translation/history [get]
func (r *V1) history(ctx fiber.Ctx) error {
	userID, ok := ctx.Locals("userID").(string)
	if !ok {
		return errorResponse(ctx, http.StatusUnauthorized, "unauthorized")
	}

	translationHistory, err := r.t.History(ctx.Context(), userID)
	if err != nil {
		r.l.Error(err, "restapi - v1 - history")

		return errorResponse(ctx, http.StatusInternalServerError, "database problems")
	}

	return okResponse(ctx, http.StatusOK, translationHistory)
}

// @Summary     Translate
// @Description Translate a text
// @ID          do-translate
// @Tags        translation
// @Accept      json
// @Produce     json
// @Param       request body     request.Translate true "Set up translation"
// @Success     200     {object} response.Ok[entity.Translation]{Code=int,Message=string,Data=entity.Translation}
// @Failure     400     {object} response.Error
// @Failure     401     {object} response.Error
// @Failure     500     {object} response.Error
// @Security    BearerAuth
// @Router      /translation/do-translate [post]
func (r *V1) doTranslate(ctx fiber.Ctx) error {
	userID, ok := ctx.Locals("userID").(string)
	if !ok {
		return errorResponse(ctx, http.StatusUnauthorized, "unauthorized")
	}

	var body request.Translate

	if err := ctx.Bind().Body(&body); err != nil {
		r.l.Error(err, "restapi - v1 - doTranslate")

		return errorResponse(ctx, http.StatusBadRequest, "invalid request body")
	}

	if err := r.v.Struct(body); err != nil {
		r.l.Error(err, "restapi - v1 - doTranslate")

		return errorResponse(ctx, http.StatusBadRequest, "invalid request body")
	}

	translation, err := r.t.Translate(
		ctx.Context(),
		userID,
		entity.Translation{
			Source:      body.Source,
			Destination: body.Destination,
			Original:    body.Original,
		},
	)
	if err != nil {
		r.l.Error(err, "restapi - v1 - doTranslate")

		return errorResponse(ctx, http.StatusInternalServerError, "translation service problems")
	}

	return okResponse(ctx, http.StatusOK, translation)
}